chore: improve workflow (#455)
### Description

<!--
Summary of change.
Example: Add sepolia chain
-->

Improve the workflow a bit with separating validation and optimization:
- Separate SVG validation and optimization
- Refactor files by making the access to get files path easier
- Add workflow for validation of files 

PD: still need to change optimization workflow to run on (probably)
every commit to main

### Backward compatibility

<!--
Are these changes backward compatible? Note that additions are backwards
compatible.

Yes/No
-->
Yes

### Testing

<!--
Have any new metadata configs and deployment addresses been used with
any Hyperlane tooling, such as the CLI?
-->
Manual
