+++ .changeset/fresh-turkeys-grow.md
@@ -1,0 +1,5 @@
+---
+'@hyperlane-xyz/registry': patch
+---
+
+Improve CI workflow by separating concerns
