+++ .github/workflows/optimize-svg.yml
@@ -32,7 +32,7 @@ jobs:
           fi
 
       - name: setup-node
-        uses: actions/setup-node@v3
+        uses: actions/setup-node@v4
         with:
           node-version: 20
