+++ .github/workflows/validate.yaml
@@ -1,0 +1,45 @@
+name: validate-files
+
+on:
+  push:
+    branches: ["main"]
+  pull_request:
+    # Allows you to run this workflow manually
+  workflow_dispatch:
+
+jobs:
+  validate-files:
+    runs-on: ubuntu-latest
+    permissions:
+      contents: read
+    steps:
+      - uses: actions/checkout@v4
+      - uses: actions/cache@v4
+        with:
+          path: |
+            **/node_modules
+            .yarn/cache
+          key: ${{ runner.os }}-yarn-cache-${{ hashFiles('./yarn.lock') }}
+
+      - name: yarn-install
+        run: |
+          yarn install
+          CHANGES=$(git status -s)
+          if [[ ! -z $CHANGES ]]; then
+            echo "Changes found: $CHANGES"
+            git diff
+            exit 1
+          fi
+
+      - name: setup-node
+        uses: actions/setup-node@v4
+        with:
+          node-version: 20
+
+      - name: validate-file-path
+        run: |
+          node ./scripts/validate-file-path.js
+
+      - name: validate-svg
+        run: |
+          node ./scripts/validate-svg.js
